refactor: reorganize BN254 crypto module structure
Move BN254 G1/G2 point operations from crypto/bls/bn254_point.ts to crypto/bn254/index.ts following the pattern established by the Grumpkin module. This creates a dedicated module for BN254 operations that can be expanded with barretenberg bbapi operations in the future.

Changes:
- Create new crypto/bn254/index.ts with Bn254 class
- Maintain backwards compatibility via re-exports in crypto/bls/index.ts
- Update tests to import from new location
- Remove old crypto/bls/bn254_point.ts

The new structure uses @noble/curves for now, with a clear path to migrate to barretenberg bbapi operations once they are implemented in bbapi_ecc.hpp.
